backend/pyproject.toml
[tool.poetry]
name = "reactor"
version = "0.1.0"
description = ""
authors = ["Bachynin Ivan <bachynin.i@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2"
pytest = "^4.5"
pytest-django = "^3.4"
pytest-mock = "^1.10"
django_extensions = "^2.1"
psycopg2-binary = "^2.8"
redis = "^3.2"
python-telegram-bot = "=12.0.0b1"
emoji = "^0.5.2"
regex = "^2019.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
